move custom_model folder to gluonts_forecasts folder
